Move replaceColumns function from controller to model
