+++ models/SecurityEventsPage.php
@@ -3,6 +3,7 @@
 namespace app\models;
 
 use Yii;
+use yii\helpers\ArrayHelper;
 
 /**
  * This is the model class for table "security_events_page".
@@ -90,5 +91,69 @@ class SecurityEventsPage extends \yii\db\ActiveRecord
     public function getUser()
     {
         return $this->hasOne(User::class, ['id' => 'user_id']);
+    }
+
+    public static function replaceColumns($rawDataColumns, $searchModel)
+    {
+        $replaceDataColumns = [];
+        array_push($replaceDataColumns, ['class' => 'yii\grid\SerialColumn',]);
+
+        foreach($rawDataColumns as $column)
+        {
+            switch ($column) {
+                case 'cef_severity':
+                    array_push($replaceDataColumns, [
+                        'attribute' => 'cef_severity',
+                        'value' => 'cef_severity',
+                        'contentOptions' => function ($dataProvider, $key, $index, $column) {
+                            $array = [
+                                ['id' => '1', 'data' => '#00DBFF'],
+                                ['id' => '2', 'data' => '#00DBFF'],
+                                ['id' => '3', 'data' => '#00FF00'],
+                                ['id' => '4', 'data' => '#00FF00'],
+                                ['id' => '5', 'data' => '#FFFF00'],
+                                ['id' => '6', 'data' => '#FFFF00'],
+                                ['id' => '7', 'data' => '#CC5500'],
+                                ['id' => '8', 'data' => '#CC5500'],
+                                ['id' => '9', 'data' => '#FF0000'],
+                                ['id' => '10', 'data' => '#FF0000'],
+                            ];
+                            if (0 < $dataProvider->cef_severity && $dataProvider->cef_severity < 11){
+                                $map = ArrayHelper::map($array, 'id', 'data');
+                                return ['style' => 'background-color:'.$map[$dataProvider->cef_severity]];
+                            } else {
+                                return ['style' => 'background-color:#FFFFFF'];
+                            }
+                        }
+                    ]);
+                    break;
+                case 'datetime':
+                    array_push($replaceDataColumns, [
+                        'attribute' => 'datetime',
+                        'value' => 'datetime',
+                        'format' => 'raw',
+                        'filter' => \macgyer\yii2materializecss\widgets\form\DatePicker::widget([
+                            'model' => $searchModel,
+                            'attribute' => 'datetime',
+                            'clientOptions' => [
+                                'format' => 'yyyy-mm-dd'
+                            ]
+                        ])
+                    ]);
+                    break;
+                case 'analyzed':
+                    array_push($replaceDataColumns, [
+                        'class' => '\dosamigos\grid\columns\BooleanColumn',
+                        'attribute' => 'analyzed',
+                        'treatEmptyAsFalse' => true
+                    ]);
+                    break;
+                default:
+                    array_push($replaceDataColumns, $column);
+            }
+        }
+        array_push($replaceDataColumns, ['class' => 'macgyer\yii2materializecss\widgets\grid\ActionColumn', 'template'=>'{view}']);
+
+        return $replaceDataColumns;
     }
 }
